Move `clearSelection` function to SidebarContentController
This function is no longer inherited from the parent scope since
`<sidebar-content>` is now a component.

Since this helper function is only used by the sidebar, move it
to the SidebarContentController controller.
